requirements.txt
django==4.0.1
djangorestframework==3.13.1
django-filter==21.1
